requirements.txt
Flask==3.0.0
yfinance==0.2.32
finnhub-python==2.4.19
python-docx==1.1.0
requests==2.31.0
gunicorn==21.2.0